requirements.txt
pytest-playwright==0.4.3

Flask==3.0.0
Flask-Cors==4.0.0
